+++ wcfsetup/install/files/lib/form/AvatarEditForm.class.php
@@ -24,11 +24,6 @@ class AvatarEditForm extends AbstractForm {
 	/**
 	 * @inheritDoc
 	 */
-	public $enableTracking = true;
-	
-	/**
-	 * @inheritDoc
-	 */
 	public $loginRequired = true;
 	
 	/**