+++ wcfsetup/install/files/lib/form/SettingsForm.class.php
@@ -28,11 +28,6 @@ class SettingsForm extends AbstractForm {
 	/**
 	 * @inheritDoc
 	 */
-	public $enableTracking = true;
-	
-	/**
-	 * @inheritDoc
-	 */
 	public $loginRequired = true;
 	
 	/**