+++ wcfsetup/install/files/lib/form/SignatureEditForm.class.php
@@ -21,11 +21,6 @@ class SignatureEditForm extends MessageForm {
 	/**
 	 * @inheritDoc
 	 */
-	public $enableTracking = true;
-	
-	/**
-	 * @inheritDoc
-	 */
 	public $loginRequired = true;
 	
 	/**