+++ wcfsetup/install/files/lib/page/TeamPage.class.php
@@ -30,11 +30,6 @@ class TeamPage extends MultipleLinkPage {
 	/**
 	 * @inheritDoc
 	 */
-	public $enableTracking = true;
-	
-	/**
-	 * @inheritDoc
-	 */
 	public $itemsPerPage = 1000;
 	
 	/**